move remaining smoketest files out of samples folder (this branch was open a long time ago and the new files were missed in a rebase)
add remaining KV smoke tests back in

add to requirements

allow for other versions of python 3.7

update to use older interfaces

fix up smoke tests for september 2019 preview

name the matrix entry properly

do not wait for cert creation in b3
