+++ common/smoketest/key_vault_certificates.py
@@ -4,9 +4,8 @@
 # ------------------------------------
 import os
 import uuid
-from azure.identity.aio import DefaultAzureCredential
-from azure.keyvault.certificates import CertificatePolicy
-from azure.keyvault.certificates.aio import CertificateClient
+from azure.identity import DefaultAzureCredential
+from azure.keyvault.certificates import CertificateClient, CertificatePolicy
 
 
 class KeyVaultCertificates:
@@ -22,21 +21,22 @@ class KeyVaultCertificates:
 
         self.certificate_name = "cert-name-" + uuid.uuid1().hex
 
-    async def create_certificate(self):
-        await self.certificate_client.create_certificate(name=self.certificate_name, policy=CertificatePolicy.get_default())
+    def create_certificate(self):
+        print("creating certificate...")
+        self.certificate_client.create_certificate(name=self.certificate_name)
         print("\tdone")
 
-    async def get_certificate(self):
+    def get_certificate(self):
         print("Getting a certificate...")
-        certificate = await self.certificate_client.get_certificate(name=self.certificate_name)
-        print(f"\tdone, certificate: {certificate.name}.")
+        certificate = self.certificate_client.get_certificate_with_policy(name=self.certificate_name)
+        print("\tdone, certificate: %s." % certificate.name)
 
-    async def delete_certificate(self):
+    def delete_certificate(self):
         print("Deleting a certificate...")
-        deleted_certificate = await self.certificate_client.delete_certificate(name=self.certificate_name)
+        deleted_certificate = self.certificate_client.delete_certificate(name=self.certificate_name)
         print("\tdone: " + deleted_certificate.name)
 
-    async def run(self):
+    def run(self):
         print("")
         print("------------------------")
         print("Key Vault - Certificates\nIdentity - Credential")
@@ -47,7 +47,7 @@ class KeyVaultCertificates:
         print("")
 
         try:
-            await self.create_certificate()
-            await self.get_certificate()
+            self.create_certificate()
+            self.get_certificate()
         finally:
-            await self.delete_certificate()
+            self.delete_certificate()
