+++ common/smoketest/key_vault_keys_async.py
@@ -4,8 +4,8 @@
 # ------------------------------------
 import os
 import uuid
-from azure.identity import DefaultAzureCredential
-from azure.keyvault.keys import KeyClient
+from azure.identity.aio import DefaultAzureCredential
+from azure.keyvault.keys.aio import KeyClient
 
 
 class KeyVaultKeys:
@@ -21,22 +21,23 @@ class KeyVaultKeys:
 
         self.key_name = "key-name-" + uuid.uuid1().hex
 
-    def create_rsa_key(self):
+    async def create_rsa_key(self):
         print("Creating an RSA key...")
-        self.key_client.create_rsa_key(name=self.key_name, size=2048)
+        await self.key_client.create_rsa_key(name=self.key_name, size=2048, hsm=False)
         print("\tdone")
 
-    def get_key(self):
+    async def get_key(self):
         print("Getting a key...")
-        key = self.key_client.get_key(name=self.key_name)
+        key = await self.key_client.get_key(name=self.key_name)
         print(f"\tdone, key: {key.name}.")
 
-    def delete_key(self):
+    async def delete_key(self):
         print("Deleting a key...")
-        deleted_key = self.key_client.begin_delete_key(name=self.key_name).result()
+        deleted_key = await self.key_client.delete_key(name=self.key_name)
         print("\tdone: " + deleted_key.name)
 
-    def run(self):
+    async def run(self):
+
         print("")
         print("------------------------")
         print("Key Vault - Keys\nIdentity - Credential")
@@ -47,7 +48,7 @@ class KeyVaultKeys:
         print("")
 
         try:
-            self.create_rsa_key()
-            self.get_key()
+            await self.create_rsa_key()
+            await self.get_key()
         finally:
-            self.delete_key()
+            await self.delete_key()
