+++ common/smoketest/program.py
@@ -2,7 +2,11 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
+import logging
+logging.basicConfig()
+
 import smoke_test
+
 
 try:
     import smoke_test_async
