+++ common/smoketest/requirements.txt
@@ -3,6 +3,8 @@ azure-core==1.0.0b3
 azure-cosmos==4.0.0b2
 azure-eventhub==5.0.0b3
 azure-identity==1.0.0b3
+azure-keyvault-certificates==4.0.0b3
+azure-keyvault-keys==4.0.0b3
 azure-keyvault-secrets==4.0.0b3
 azure-storage-blob==12.0.0b3
 azure-storage-common==2.1.0
