+++ common/smoketest/smoke_test.py
@@ -2,6 +2,8 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
+from key_vault_certificates import KeyVaultCertificates
+from key_vault_keys import KeyVaultKeys
 from key_vault_secrets import KeyVaultSecrets
 from storage_blob import StorageBlob
 from event_hubs import EventHub
@@ -12,6 +14,8 @@ print("==========================================")
 print("      AZURE TRACK 2 SDKs SMOKE TEST")
 print("==========================================")
 
+KeyVaultCertificates().run()
+KeyVaultKeys().run()
 KeyVaultSecrets().run()
 StorageBlob().run()
 EventHub().run()
