+++ common/smoketest/smoke_test_async.py
@@ -3,6 +3,8 @@
 # Licensed under the MIT License.
 # ------------------------------------
 import asyncio
+from key_vault_certificates_async import KeyVaultCertificates
+from key_vault_keys_async import KeyVaultKeys
 from key_vault_secrets_async import KeyVaultSecrets
 from event_hubs_async import EventHub
 
@@ -13,6 +15,8 @@ print("==========================================")
 
 
 async def main():
+    await KeyVaultCertificates().run()
+    await KeyVaultKeys().run()
     await KeyVaultSecrets().run()
     await EventHub().run()
